// src/integration/api-types.ts
export interface components {
    schemas: {
// ...
        Begrip: {
            id: string;
            /**
             * @description skos:prefLabel
             * @example Applicatiefunctie
             */
            term: string;
            /**
             * @description skos:definition
             * @example Een functie van een applicatie
             */
            definitie: string;
            /** @description rdfs:comment */
            toelichting?: string | null;
            /** @description skos:scopeNote */
            scopeNote?: string | null;
            /** @description Bronverwijzing */
            bron?: string | null;
            /**
             * @description NL-SBB concept URI
             * @example https://begrippen.noraonline.nl/basisbegrippen/applicatiefunctie
             */
            uri: string;
            /**
             * @description skos:altLabel
             * @example [
             *       "appfunctie"
             *     ]
             */
            synoniemen: string[];
            /**
             * @description SKOSMOS vocabulary ID
             * @example basisbegrippen
             */
            vocab: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}